TOB-5263 implementere logikk for utsending til Arena

+++ src/main/kotlin/no/nav/sokos/skattekort/module/utsending/oppdragz/Skattekort.kt
@@ -3,7 +3,7 @@ package no.nav.sokos.skattekort.module.utsending.oppdragz
 import javax.xml.datatype.XMLGregorianCalendar
 
 data class Skattekort(
-    val inntektsaar: Long,
+    val inntektsaar: Int,
     val utstedtDato: XMLGregorianCalendar,
     val skattekortidentifikator: Long,
     val forskuddstrekk: List<Forskuddstrekk> = listOf(),
